api/requirements.txt
-r ../requirements.txt
Beaker == 1.6.4
#Jinja2==2.9
Pillow == 3.3.0
PyYAML == 3.10
chill==0.4.0
docopt
gevent-websocket
html == 1.16
piecemaker == 0.2.4
progressbar == 2.3
pycurl
pysqlite
redis
rq
websocket-client
requests
Flask-Sockets
simplejson == 3.1.3
flask-secure-cookie == 0.1.2
cairosvg==1.0.22
pycairo
setuptools_scm
-e api/
